refactor: change directory for pages and ui component
