+++ src/components/Layout/Layout.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
-import Header from '../Header/Header';
+import Header from '../ui/Header/Header';
 
 function Layout() {
   return (
