+++ src/components/pages/index.tsx
@@ -1,0 +1,4 @@
+export { default as FormPage } from './FormPage/FormPage';
+export { default as AboutPage } from './AboutPage/AboutPage';
+export { default as ErrorPage } from './ErrorPage/ErrorPage';
+export { default as HomePage } from './HomePage/HomePage';
