+++ src/components/ui/ImageCardContainer/ImageCard/ImageCardPopup/ImageCardPopup.scss
