+++ src/components/ui/OrderingForm/OrderCard/OrderCard.scss
