+++ src/components/ui/OrderingForm/OrderCreatedWindow/OrderCreatedWindow.scss
