+++ src/components/ui/SearchBar/SearchBar.scss
